error fixed from dak

+++ src/dNothi.Desktop/UI/GuardFileUI/GuardFileUserControls/GuardFileTypeTableUserControl.cs
@@ -1,32 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using dNothi.Services.DakServices;
-using dNothi.Services.UserServices;
-using dNothi.Desktop.UI.CustomMessageBox;
-using dNothi.Services.GuardFile.Model;
+﻿using dNothi.Desktop.UI.CustomMessageBox;
 using dNothi.Services.GuardFile;
-using dNothi.Desktop.UI.GuardFileUI;
+using dNothi.Services.GuardFile.Model;
+using dNothi.Services.UserServices;
 using dNothi.Utility;
+using System;
+using System.Windows.Forms;
 
 namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
 {
     public partial class GuardFileTypeTableUserControl : UserControl
     {
-       
+
         IGuardFileService<GuardFileCategory, GuardFileCategory.Record> _guardFileService { get; set; }
-        public  const string GuardFileCategory = "GuardFileCategories";
+        public const string GuardFileCategory = "GuardFileCategories";
         IUserService _userService { get; set; }
         AllAlartMessage alartMessage = new AllAlartMessage();
         public GuardFileTypeTableUserControl(IUserService userService, IGuardFileService<GuardFileCategory, GuardFileCategory.Record> guardFileService)
         {
-         
+
             _userService = userService;
             _guardFileService = guardFileService;
             InitializeComponent();
@@ -40,20 +31,22 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
             MyToolTip.SetToolTip(saveEditButton, "সংরক্ষণ করুন");
             MyToolTip.SetToolTip(cancelButton, "বাতিল করুন");
         }
+        private bool _isOnline { get; set; }
+        public bool isOnline { get => _isOnline; set { _isOnline = value; label3.Visible = value; } }
         public int _office_unit_organogram_id { get; set; }
         public int designation_id { get; set; }
         public int TypeId { get; set; }
         public int _id { get; set; }
-       
+
         public string _decisision { get; set; }
-       
+
         public string decision
         {
             get { return _decisision; }
             set
             {
                 _decisision = value;
-               
+
                 decisionNameLabel.Text = value;
                 decisionNameTextBox.Text = value;
 
@@ -88,29 +81,35 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
 
             }
         }
-        public int id { get { return _id; } set { _id = value;label1.Text =ConversionMethod.EnglishNumberToBangla( value.ToString());  } }
+        public int id { get { return _id; } set { _id = value; label1.Text = ConversionMethod.EnglishNumberToBangla(value.ToString()); } }
 
         public string _typeNo
         {
             get; set;
         }
-        public string TypeNo { get { return _typeNo; } set { _typeNo = value;
+        public string TypeNo
+        {
+            get { return _typeNo; }
+            set
+            {
+                _typeNo = value;
                 if (Convert.ToInt32(value) > 0)
                 { label2.Text = ConversionMethod.EnglishNumberToBangla(value); }
-            } }
+            }
+        }
 
-       
+
         public event EventHandler DeleteButtonClick;
         private void decisionRadioButton_CheckedChanged(object sender, EventArgs e)
         {
-          
+
         }
 
         private void decisionEditRightButton_Click(object sender, EventArgs e)
         {
-          
 
-             EditMode();
+
+            EditMode();
         }
 
         private void EditMode()
@@ -152,7 +151,7 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
         {
             if (TypeId > 0)
             {
-               
+
                 var dakListUserParam = _userService.GetLocalDakUserParam();
 
                 GuardFileCategory.Record model = new GuardFileCategory.Record();
@@ -162,7 +161,7 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
 
 
                 var response = _guardFileService.Insert(dakListUserParam, 3, GuardFileCategory, model);
-                if(response.status=="success")
+                if (response.status == "success")
                 {
                     alartMessage.SuccessMessage("ধরন সংরক্ষণ সফল হয়েছে।");
 
@@ -173,7 +172,7 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
                     alartMessage.ErrorMessage("পুনরায় চেষ্ঠা করুন।");
                 }
             }
-            
+
         }
 
         private void decisionDeleteButton_Click(object sender, EventArgs e)
@@ -182,13 +181,14 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
             conditonBoxForm.message = "আপনি কি নিশ্চিতভাবে সিদ্ধান্ত টি মুছে ফেলতে চান?";
             conditonBoxForm.ShowDialog();
             if (conditonBoxForm.Yes)
-            { if(this.decisionDeleteButton!=null)
+            {
+                if (this.decisionDeleteButton != null)
                 {
-                    DeleteButtonClick(sender,e);
+                    DeleteButtonClick(sender, e);
                 }
-               
+
             }
-       
+
 
         }
 
@@ -211,12 +211,12 @@ namespace dNothi.Desktop.UI.OtherModule.GuardFileUserControls
         {
             UIDesignCommonMethod.Table_Color_Blue(sender, e);
         }
-       
+
         public event EventHandler GuardFileCountLabelClick;
         private void label2_Click(object sender, EventArgs e)
         {
             if (this.GuardFileCountLabelClick != null)
-                    this.GuardFileCountLabelClick(sender, e);
+                this.GuardFileCountLabelClick(sender, e);
         }
 
         //Update
